+++ pyvisim/encoders/utils.py
@@ -1,1 +1,0 @@
-from ..utils.image_utils import iter_images